almost finished

+++ ft_libft/LIBFT/ft_strncmp.c
@@ -1,11 +1,24 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_strncmp.c                                       :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: hdal-mol <hdal-mol@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/03/18 23:38:24 by hdal-mol          #+#    #+#             */
+/*   Updated: 2021/03/19 18:15:24 by hdal-mol         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 
 
 /*
 ** compara as  strings s1 and s2 char-wise, returning the difference.
 */
+#include "libft.h"
 
 
-int ft_strncmp(char *s1, char *s2, int n)
+int ft_strncmp(char *s1, char *s2, size_t n)
 {
     int i;
